Fix display CSS style
`unset` is effectively `inherit` or `initial`. In this case it's the
latter, which evaluates to `inline`. I didn't notice this until playing
around with a grid layout for a new feature, but it is making a slight
difference as is - the gap between this and the next paragraph

`revert` meanwhile resets the `display` property to what it would be
without `none` having been applied (this is a simplication, but it's
true in this case), which in this case is `block`

https://developer.mozilla.org/en-US/docs/Web/CSS/unset
https://developer.mozilla.org/en-US/docs/Web/CSS/revert